Rename reminder_frequency file
