+++ lib/reminders/domain/reminder.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:petto/reminders/domain/reminder_frecuency.dart';
+import 'package:petto/reminders/domain/reminder_frequency.dart';
 import 'package:petto/reminders/domain/reminder_type.dart';
 
 part 'reminder.freezed.dart';
